+++ tests/ClientTest.php
@@ -35,4 +35,49 @@ class ClientTest extends \PHPUnit\Framework\TestCase
             ['/dir/', 200, null],
         ];
     }
+
+    /**
+     * @dataProvider uploadDataProvider
+     */
+    public function testUpload(string $file, bool $withChecksum, ?string $expectedChecksum)
+    {
+        $options = function (array $options) use ($expectedChecksum): bool {
+            if (!is_resource($options['body'])) {
+                return false;
+            }
+
+            if (null === $expectedChecksum) {
+                return !isset($options['headers']['Checksum']);
+            }
+
+            if (!isset($options['headers']['Checksum'])) {
+                return false;
+            }
+
+            if ($expectedChecksum !== $options['headers']['Checksum']) {
+                return false;
+            }
+
+            return true;
+        };
+
+        $response = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
+        $response->expects($this->atLeastOnce())->method('getStatusCode')->willReturn(201);
+
+        $httpClient = $this->createMock(\GuzzleHttp\Client::class);
+        $httpClient->expects($this->once())->method('request')->with('PUT', 'test/'.$file, $this->callback($options))->willReturn($response);
+
+        $client = new Client('abc1234d', 'test', Region::FALKENSTEIN, $httpClient);
+        $client->upload(__DIR__.'/_files/'.$file, $file, $withChecksum);
+    }
+
+    public static function uploadDataProvider(): array
+    {
+        return [
+            ['a.txt', true, 'ECD71870D1963316A97E3AC3408C9835AD8CF0F3C1BC703527C30265534F75AE'],
+            ['a.txt', false, null],
+            ['bunny-logo.jpg', true, '3A2CBCFCA2CBF58B842B15B08AC69FEC706284CD72CC48058840B4E448AE5949'],
+            ['bunny-logo.jpg', false, null],
+        ];
+    }
 }