+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Checksum uploads in `Client::upload()` and `Client::putContents()`. It can be disabled using the `bool $withChecksum` parameter;
+
 ## [3.1.0] - 2024-03-25
 
 ### Added